+++ src-input/duk_hobject_alloc.c
@@ -12,18 +12,23 @@
 
 #include "duk_internal.h"
 
-DUK_LOCAL void duk__init_object_parts(duk_heap *heap, duk_hobject *obj, duk_uint_t hobject_flags) {
+/*
+ *  Helpers.
+ */
+
+DUK_LOCAL void duk__init_object_parts(duk_heap *heap, duk_uint_t hobject_flags, duk_hobject *obj) {
+	DUK_ASSERT(obj != NULL);
+	/* Zeroed by caller. */
+
+	obj->hdr.h_flags = hobject_flags | DUK_HTYPE_OBJECT;
+	DUK_ASSERT(DUK_HEAPHDR_GET_TYPE(&obj->hdr) == DUK_HTYPE_OBJECT);  /* Assume zero shift. */
+
 #if defined(DUK_USE_EXPLICIT_NULL_INIT)
 	DUK_HOBJECT_SET_PROTOTYPE(heap, obj, NULL);
 	DUK_HOBJECT_SET_PROPS(heap, obj, NULL);
 #endif
-
-	/* XXX: macro? sets both heaphdr and object flags */
-	obj->hdr.h_flags = hobject_flags;
-	DUK_HEAPHDR_SET_TYPE(&obj->hdr, DUK_HTYPE_OBJECT);  /* also goes into flags */
-
 #if defined(DUK_USE_HEAPPTR16)
-	/* Zero encoded pointer is required to match NULL */
+	/* Zero encoded pointer is required to match NULL. */
 	DUK_HEAPHDR_SET_NEXT(heap, &obj->hdr, NULL);
 #if defined(DUK_USE_DOUBLE_LINKED_HEAP)
 	DUK_HEAPHDR_SET_PREV(heap, &obj->hdr, NULL);
@@ -32,12 +37,20 @@ DUK_LOCAL void duk__init_object_parts(duk_heap *heap, duk_hobject *obj, duk_uint
 	DUK_ASSERT_HEAPHDR_LINKS(heap, &obj->hdr);
 	DUK_HEAP_INSERT_INTO_HEAP_ALLOCATED(heap, &obj->hdr);
 
-	/*
-	 *  obj->props is intentionally left as NULL, and duk_hobject_props.c must deal
-	 *  with this properly.  This is intentional: empty objects consume a minimum
-	 *  amount of memory.  Further, an initial allocation might fail and cause
-	 *  'obj' to "leak" (require a mark-and-sweep) since it is not reachable yet.
+	/* obj->props is intentionally left as NULL, and duk_hobject_props.c must deal
+	 * with this properly.  This is intentional: empty objects consume a minimum
+	 * amount of memory.  Further, an initial allocation might fail and cause
+	 * 'obj' to "leak" (require a mark-and-sweep) since it is not reachable yet.
 	 */
+}
+
+DUK_LOCAL void *duk__hobject_alloc_init(duk_hthread *thr, duk_uint_t hobject_flags, duk_size_t size) {
+	void *res;
+
+	res = (void *) DUK_ALLOC_CHECKED_ZEROED(thr, size);
+	DUK_ASSERT(res != NULL);
+	duk__init_object_parts(thr->heap, hobject_flags, (duk_hobject *) res);
+	return res;
 }
 
 /*
@@ -51,7 +64,7 @@ DUK_LOCAL void duk__init_object_parts(duk_heap *heap, duk_hobject *obj, duk_uint
  *  count before invoking any operation that might require memory allocation.
  */
 
-DUK_INTERNAL duk_hobject *duk_hobject_alloc(duk_heap *heap, duk_uint_t hobject_flags) {
+DUK_INTERNAL duk_hobject *duk_hobject_alloc_unchecked(duk_heap *heap, duk_uint_t hobject_flags) {
 	duk_hobject *res;
 
 	DUK_ASSERT(heap != NULL);
@@ -61,28 +74,27 @@ DUK_INTERNAL duk_hobject *duk_hobject_alloc(duk_heap *heap, duk_uint_t hobject_f
 	DUK_ASSERT((hobject_flags & DUK_HOBJECT_FLAG_NATFUNC) == 0);
 	DUK_ASSERT((hobject_flags & DUK_HOBJECT_FLAG_THREAD) == 0);
 
-	res = (duk_hobject *) DUK_ALLOC(heap, sizeof(duk_hobject));
+	res = (duk_hobject *) DUK_ALLOC_ZEROED(heap, sizeof(duk_hobject));
 	if (DUK_UNLIKELY(res == NULL)) {
 		return NULL;
 	}
-	DUK_MEMZERO(res, sizeof(duk_hobject));
 
-	duk__init_object_parts(heap, res, hobject_flags);
+	duk__init_object_parts(heap, hobject_flags, res);
 
 	return res;
 }
 
-DUK_INTERNAL duk_hcompfunc *duk_hcompfunc_alloc(duk_heap *heap, duk_uint_t hobject_flags) {
+DUK_INTERNAL duk_hobject *duk_hobject_alloc(duk_hthread *thr, duk_uint_t hobject_flags) {
+	duk_hobject *res;
+
+	res = (duk_hobject *) duk__hobject_alloc_init(thr, hobject_flags, sizeof(duk_hobject));
+	return res;
+}
+
+DUK_INTERNAL duk_hcompfunc *duk_hcompfunc_alloc(duk_hthread *thr, duk_uint_t hobject_flags) {
 	duk_hcompfunc *res;
 
-	res = (duk_hcompfunc *) DUK_ALLOC(heap, sizeof(duk_hcompfunc));
-	if (DUK_UNLIKELY(res == NULL)) {
-		return NULL;
-	}
-	DUK_MEMZERO(res, sizeof(duk_hcompfunc));
-
-	duk__init_object_parts(heap, &res->obj, hobject_flags);
-
+	res = (duk_hcompfunc *) duk__hobject_alloc_init(thr, hobject_flags, sizeof(duk_hcompfunc));
 #if defined(DUK_USE_EXPLICIT_NULL_INIT)
 #if defined(DUK_USE_HEAPPTR16)
 	/* NULL pointer is required to encode to zero, so memset is enough. */
@@ -98,17 +110,10 @@ DUK_INTERNAL duk_hcompfunc *duk_hcompfunc_alloc(duk_heap *heap, duk_uint_t hobje
 	return res;
 }
 
-DUK_INTERNAL duk_hnatfunc *duk_hnatfunc_alloc(duk_heap *heap, duk_uint_t hobject_flags) {
+DUK_INTERNAL duk_hnatfunc *duk_hnatfunc_alloc(duk_hthread *thr, duk_uint_t hobject_flags) {
 	duk_hnatfunc *res;
 
-	res = (duk_hnatfunc *) DUK_ALLOC(heap, sizeof(duk_hnatfunc));
-	if (DUK_UNLIKELY(res == NULL)) {
-		return NULL;
-	}
-	DUK_MEMZERO(res, sizeof(duk_hnatfunc));
-
-	duk__init_object_parts(heap, &res->obj, hobject_flags);
-
+	res = (duk_hnatfunc *) duk__hobject_alloc_init(thr, hobject_flags, sizeof(duk_hnatfunc));
 #if defined(DUK_USE_EXPLICIT_NULL_INIT)
 	res->func = NULL;
 #endif
@@ -117,17 +122,10 @@ DUK_INTERNAL duk_hnatfunc *duk_hnatfunc_alloc(duk_heap *heap, duk_uint_t hobject
 }
 
 #if defined(DUK_USE_BUFFEROBJECT_SUPPORT)
-DUK_INTERNAL duk_hbufobj *duk_hbufobj_alloc(duk_heap *heap, duk_uint_t hobject_flags) {
+DUK_INTERNAL duk_hbufobj *duk_hbufobj_alloc(duk_hthread *thr, duk_uint_t hobject_flags) {
 	duk_hbufobj *res;
 
-	res = (duk_hbufobj *) DUK_ALLOC(heap, sizeof(duk_hbufobj));
-	if (DUK_UNLIKELY(res == NULL)) {
-		return NULL;
-	}
-	DUK_MEMZERO(res, sizeof(duk_hbufobj));
-
-	duk__init_object_parts(heap, &res->obj, hobject_flags);
-
+	res = (duk_hbufobj *) duk__hobject_alloc_init(thr, hobject_flags, sizeof(duk_hbufobj));
 #if defined(DUK_USE_EXPLICIT_NULL_INIT)
 	res->buf = NULL;
 	res->buf_prop = NULL;
@@ -138,15 +136,13 @@ DUK_INTERNAL duk_hbufobj *duk_hbufobj_alloc(duk_heap *heap, duk_uint_t hobject_f
 }
 #endif  /* DUK_USE_BUFFEROBJECT_SUPPORT */
 
-/*
- *  Allocate a new thread.
+/* Allocate a new thread.
  *
- *  Leaves the built-ins array uninitialized.  The caller must either
- *  initialize a new global context or share existing built-ins from
- *  another thread.
+ * Leaves the built-ins array uninitialized.  The caller must either
+ * initialize a new global context or share existing built-ins from
+ * another thread.
  */
-
-DUK_INTERNAL duk_hthread *duk_hthread_alloc(duk_heap *heap, duk_uint_t hobject_flags) {
+DUK_INTERNAL duk_hthread *duk_hthread_alloc_unchecked(duk_heap *heap, duk_uint_t hobject_flags) {
 	duk_hthread *res;
 
 	res = (duk_hthread *) DUK_ALLOC(heap, sizeof(duk_hthread));
@@ -155,7 +151,7 @@ DUK_INTERNAL duk_hthread *duk_hthread_alloc(duk_heap *heap, duk_uint_t hobject_f
 	}
 	DUK_MEMZERO(res, sizeof(duk_hthread));
 
-	duk__init_object_parts(heap, &res->obj, hobject_flags);
+	duk__init_object_parts(heap, hobject_flags, &res->obj);
 
 #if defined(DUK_USE_EXPLICIT_NULL_INIT)
 	res->ptr_curr_pc = NULL;
@@ -174,7 +170,7 @@ DUK_INTERNAL duk_hthread *duk_hthread_alloc(duk_heap *heap, duk_uint_t hobject_f
 	res->strs = NULL;
 #endif
 	{
-		int i;
+		duk_small_uint_t i;
 		for (i = 0; i < DUK_NUM_BUILTINS; i++) {
 			res->builtins[i] = NULL;
 		}
@@ -191,51 +187,30 @@ DUK_INTERNAL duk_hthread *duk_hthread_alloc(duk_heap *heap, duk_uint_t hobject_f
 	return res;
 }
 
-#if 0  /* unused now */
-DUK_INTERNAL duk_hobject *duk_hobject_alloc_checked(duk_hthread *thr, duk_uint_t hobject_flags) {
-	duk_hobject *res = duk_hobject_alloc(thr->heap, hobject_flags);
-	if (!res) {
+DUK_INTERNAL duk_hthread *duk_hthread_alloc(duk_hthread *thr, duk_uint_t hobject_flags) {
+	duk_hthread *res;
+
+	res = duk_hthread_alloc_unchecked(thr->heap, hobject_flags);
+	if (res == NULL) {
 		DUK_ERROR_ALLOC_FAILED(thr);
 	}
 	return res;
 }
-#endif
 
-/*
- *  Allocate a new array.
- */
-
-DUK_INTERNAL duk_harray *duk_harray_alloc(duk_heap *heap, duk_uint_t hobject_flags) {
+DUK_INTERNAL duk_harray *duk_harray_alloc(duk_hthread *thr, duk_uint_t hobject_flags) {
 	duk_harray *res;
 
-	res = (duk_harray *) DUK_ALLOC(heap, sizeof(duk_harray));
-	if (DUK_UNLIKELY(res == NULL)) {
-		return NULL;
-	}
-	DUK_MEMZERO(res, sizeof(duk_harray));
-
-	duk__init_object_parts(heap, &res->obj, hobject_flags);
+	res = (duk_harray *) duk__hobject_alloc_init(thr, hobject_flags, sizeof(duk_harray));
 
 	DUK_ASSERT(res->length == 0);
 
 	return res;
 }
 
-/*
- *  Allocate a new environment
- */
-
-DUK_INTERNAL duk_hdecenv *duk_hdecenv_alloc(duk_heap *heap, duk_uint_t hobject_flags) {
+DUK_INTERNAL duk_hdecenv *duk_hdecenv_alloc(duk_hthread *thr, duk_uint_t hobject_flags) {
 	duk_hdecenv *res;
 
-	res = (duk_hdecenv *) DUK_ALLOC(heap, sizeof(duk_hdecenv));
-	if (DUK_UNLIKELY(res == NULL)) {
-		return NULL;
-	}
-	DUK_MEMZERO(res, sizeof(duk_hdecenv));
-
-	duk__init_object_parts(heap, &res->obj, hobject_flags);
-
+	res = (duk_hdecenv *) duk__hobject_alloc_init(thr, hobject_flags, sizeof(duk_hdecenv));
 #if defined(DUK_USE_EXPLICIT_NULL_INIT)
 	res->thread = NULL;
 	res->varmap = NULL;
@@ -248,17 +223,10 @@ DUK_INTERNAL duk_hdecenv *duk_hdecenv_alloc(duk_heap *heap, duk_uint_t hobject_f
 	return res;
 }
 
-DUK_INTERNAL duk_hobjenv *duk_hobjenv_alloc(duk_heap *heap, duk_uint_t hobject_flags) {
+DUK_INTERNAL duk_hobjenv *duk_hobjenv_alloc(duk_hthread *thr, duk_uint_t hobject_flags) {
 	duk_hobjenv *res;
 
-	res = (duk_hobjenv *) DUK_ALLOC(heap, sizeof(duk_hobjenv));
-	if (DUK_UNLIKELY(res == NULL)) {
-		return NULL;
-	}
-	DUK_MEMZERO(res, sizeof(duk_hobjenv));
-
-	duk__init_object_parts(heap, &res->obj, hobject_flags);
-
+	res = (duk_hobjenv *) duk__hobject_alloc_init(thr, hobject_flags, sizeof(duk_hobjenv));
 #if defined(DUK_USE_EXPLICIT_NULL_INIT)
 	res->target = NULL;
 #endif